+++ builtin/logical/pki/path_issue_sign.go
@@ -171,8 +171,9 @@ func (b *backend) pathIssueSignCert(
 
 	resp := b.Secret(SecretCertsType).Response(
 		map[string]interface{}{
-			"certificate": cb.Certificate,
-			"issuing_ca":  cb.IssuingCA,
+			"certificate":   cb.Certificate,
+			"issuing_ca":    cb.IssuingCA,
+			"serial_number": cb.SerialNumber,
 		},
 		map[string]interface{}{
 			"serial_number": cb.SerialNumber,